ml/dl/llm/agent/examples/requirements.txt
openai>=1.37.0
anthropic>=0.34.1
cohere>=5.9.4
groq>=0.9.0
mistralai>=0.4.1
google-generativeai>=0.7.2
ai21>=2.2.0
httpx>=0.27.0
requests>=2.32.3
pydantic>=2.8.2
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
uvloop; sys_platform != 'win32'
rich>=13.7.1
click>=8.1.7
firecrawl-py>=0.0.21
duckduckgo-search>=6.2.10
langchain>=0.2.12
langgraph>=0.2.14
faiss-cpu>=1.8.0.post1
chromadb>=0.5.6
arxiv>=2.1.3
pypdf>=4.3.1
numexpr>=2.10.1
pandas>=2.2.2
numpy>=2.0.1
beautifulsoup4>=4.12.3
playwright>=1.46.0
browser-use>=0.1.25
termcolor>=2.4.0
loguru>=0.7.2
tenacity>=8.5.0
structlog>=24.1.0
fastapi>=0.112.0
uvicorn>=0.30.5
retrying>=1.3.4
sse-starlette>=2.1.3
websockets>=12.0
crewai>=0.64.1
pyautogen>=0.2.33
swarm @ git+https://github.com/openai/swarm.git
